cassandra-8773 Commit log replayer deletes files with invalid mutations after each replay cycle (#1031)

+++ src/java/org/apache/cassandra/db/commitlog/AbstractCommitLogSegmentManager.java
@@ -415,9 +415,21 @@ public abstract class AbstractCommitLogSegmentManager
      */
     void handleReplayedSegment(final File file)
     {
-        // (don't decrease managed size, since this was never a "live" segment)
-        logger.trace("(Unopened) segment {} is no longer needed and will be deleted now", file);
-        FileUtils.deleteWithConfirm(file);
+        handleReplayedSegment(file, false);
+    }
+
+    void handleReplayedSegment(final File file, boolean hasInvalidMutations)
+    {
+        if (!hasInvalidMutations)
+        {
+            // (don't decrease managed size, since this was never a "live" segment)
+            logger.trace("(Unopened) segment {} is no longer needed and will be deleted now", file);
+            FileUtils.deleteWithConfirm(file);
+        }
+        else
+        {
+            logger.debug("File {} should not be deleted as it contains invalid mutations", file.name());
+        }
     }
 
     /**
